+++ src/Resources/contao/languages/en/default.php
@@ -1,4 +1,4 @@
-<?php if (!defined('TL_ROOT')) die('You cannot access this file directly!');
+<?php
 
 /**
  * Contao Open Source CMS
@@ -11,8 +11,5 @@
  * @copyright Frank Hoppe 2013
  */
 
-$GLOBALS['TL_LANG']['CTE']['schach'] = "Chess-Elements";
-$GLOBALS['TL_LANG']['CTE']['chesstable']['0'] = "Chess table";
-$GLOBALS['TL_LANG']['CTE']['chesstable']['1'] = "Output CSV data as a chess table";
-
-?>
+$GLOBALS['TL_LANG']['CTE']['schach'] = 'Chess-Elements';
+$GLOBALS['TL_LANG']['CTE']['chesstable'] = array('Chess table', 'Output CSV data as a chess table');