+++ src/Resources/contao/languages/en/tl_settings.php
@@ -1,4 +1,4 @@
-<?php if (!defined('TL_ROOT')) die('You cannot access this file directly!');
+<?php
 
 /**
  * Contao Open Source CMS
@@ -14,7 +14,7 @@
 /**
  * Name
  */
-$GLOBALS['TL_LANG']['tl_settings']['chesstable']              = "Chess table";
+$GLOBALS['TL_LANG']['tl_settings']['chesstable']              = 'Chess table';
 
 /**
  * legends
@@ -24,27 +24,14 @@ $GLOBALS['TL_LANG']['tl_settings']['chesstable_legend']       = 'Chess table';
 /**
  * fields
  */
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_blindfelder']['0'] = "Table values ​​for blind fields";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_blindfelder']['1'] = "Separate the values ​​with commas. The uppercase and lowercase do not need to consider.";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_nationfelder']['0'] = "Column headers for nation fields";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_nationfelder']['1'] = "Separate the values ​​with commas. The uppercase and lowercase do not need to consider.";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_platzfelder']['0'] = "Column headers for square fields";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_platzfelder']['1'] = "Separate the values ​​with commas. The uppercase and lowercase do not need to consider.";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_vereinfelder']['0'] = "Column headers for club fields";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_vereinfelder']['1'] = "Separate the values ​​with commas. The uppercase and lowercase do not need to consider.";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_namenfelder']['0'] = "Column headers for name fields";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_namenfelder']['1'] = "Separate the values ​​with commas. The uppercase and lowercase do not need to consider.";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_punktefelder']['0'] = "Column headers for point fields";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_punktefelder']['1'] = "Separate the values ​​with commas. The uppercase and lowercase do not need to consider.";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_wertungfelder']['0'] = "Column headers for rating fields";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_wertungfelder']['1'] = "Separate the values ​​with commas. The uppercase and lowercase do not need to consider.";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_ratingfelder']['0'] = "Column headers for Elo fields";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_ratingfelder']['1'] = "Separate the values ​​with commas. The uppercase and lowercase do not need to consider.";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_ergebnisfelder']['0'] = "Column headers for result fields";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_ergebnisfelder']['1'] = "Separate the values ​​with commas. The uppercase and lowercase do not need to consider.";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_farbfelder']['0'] = "Column headers for color swatches. The column header content is not displayed in the frontend.";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_farbfelder']['1'] = "Separate the values ​​with commas. The uppercase and lowercase do not need to consider.";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_steuerfelder']['0'] = "Column headers for control fields. The column is not displayed in the frontend.";
-$GLOBALS['TL_LANG']['tl_settings']['chesstable_steuerfelder']['1'] = "Separate the values ​​with commas. The uppercase and lowercase do not need to consider.";
-
-?>
+$GLOBALS['TL_LANG']['tl_settings']['chesstable_blindfelder']= array('Table values ​​for blind fields', 'Separate the values ​​with commas. The uppercase and lowercase do not need to consider.');
+$GLOBALS['TL_LANG']['tl_settings']['chesstable_nationfelder']= array('Column headers for nation fields', 'Separate the values ​​with commas. The uppercase and lowercase do not need to consider.');
+$GLOBALS['TL_LANG']['tl_settings']['chesstable_platzfelder']= array('Column headers for square fields', 'Separate the values ​​with commas. The uppercase and lowercase do not need to consider.');
+$GLOBALS['TL_LANG']['tl_settings']['chesstable_vereinfelder']= array('Column headers for club fields', 'Separate the values ​​with commas. The uppercase and lowercase do not need to consider.');
+$GLOBALS['TL_LANG']['tl_settings']['chesstable_namenfelder']= array('Column headers for name fields', 'Separate the values ​​with commas. The uppercase and lowercase do not need to consider.');
+$GLOBALS['TL_LANG']['tl_settings']['chesstable_punktefelder']= array('Column headers for point fields', 'Separate the values ​​with commas. The uppercase and lowercase do not need to consider.');
+$GLOBALS['TL_LANG']['tl_settings']['chesstable_wertungfelder']= array('Column headers for rating fields', 'Separate the values ​​with commas. The uppercase and lowercase do not need to consider.');
+$GLOBALS['TL_LANG']['tl_settings']['chesstable_ratingfelder']= array('Column headers for Elo fields', 'Separate the values ​​with commas. The uppercase and lowercase do not need to consider.');
+$GLOBALS['TL_LANG']['tl_settings']['chesstable_ergebnisfelder']= array('Column headers for result fields', 'Separate the values ​​with commas. The uppercase and lowercase do not need to consider.');
+$GLOBALS['TL_LANG']['tl_settings']['chesstable_farbfelder']= array('Column headers for color swatches. The column header content is not displayed in the frontend.', 'Separate the values ​​with commas. The uppercase and lowercase do not need to consider.');
+$GLOBALS['TL_LANG']['tl_settings']['chesstable_steuerfelder']= array('Column headers for control fields. The column is not displayed in the frontend.', 'Separate the values ​​with commas. The uppercase and lowercase do not need to consider.');